# plot_data.float()

+++ autoencoder/dataset.py
@@ -39,7 +39,7 @@ class PointCloudDataset:
 
     def get_plot_data(self):
         plot_data = torch.unsqueeze(self.data[0], 0)
-        plot_data.float()
+        # plot_data.float()
         return plot_data
 
 
@@ -70,5 +70,5 @@ class PointCloudEvalDataset:
 
     def get_plot_data(self):
         plot_data = torch.unsqueeze(self.data[0], 0)
-        plot_data.float()
+        # plot_data.float()
         return plot_data
